Make User flags support long values

+++ core/src/main/java/discord4j/core/object/entity/User.java
@@ -17,6 +17,7 @@
 package discord4j.core.object.entity;
 
 import discord4j.core.GatewayDiscordClient;
+import discord4j.core.object.Embed;
 import discord4j.core.object.entity.channel.PrivateChannel;
 import discord4j.core.retriever.EntityRetrievalStrategy;
 import discord4j.core.util.EntityUtil;
@@ -31,6 +32,7 @@ import reactor.core.publisher.Mono;
 import reactor.util.annotation.Nullable;
 
 import java.util.EnumSet;
+import java.util.Locale;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -300,7 +302,7 @@ public class User implements Entity {
      * @return A {@code EnumSet} with the public flags of this user.
      */
     public EnumSet<Flag> getPublicFlags() {
-        int publicFlags = data.publicFlags().toOptional().orElse(0);
+        long publicFlags = data.publicFlags().toOptional().orElse(0L);
         if (publicFlags != 0) {
             return Flag.of(publicFlags);
         }
@@ -388,7 +390,7 @@ public class User implements Entity {
          * @param value The flags value as represented by Discord.
          * @return The {@link EnumSet} of flags.
          */
-        public static EnumSet<Flag> of(final int value) {
+        public static EnumSet<Flag> of(final long value) {
             final EnumSet<Flag> userFlags = EnumSet.noneOf(Flag.class);
             for (Flag flag : Flag.values()) {
                 long flagValue = flag.getFlag();
